Adding tests for internal trtri interface (#1828)
[ROCm/rocblas commit: 650320edd91f62fe793f20bc7bec37d0764df08c]

+++ clients/include/blas3/testing_trtri_strided_batched.hpp
@@ -147,7 +147,7 @@ void testing_trtri_strided_batched(const Arguments& arg)
     rocblas_int lda         = arg.lda;
     rocblas_int batch_count = arg.batch_count;
 
-    rocblas_stride stride_A = lda * N;
+    rocblas_stride stride_A = size_t(lda) * N;
 
     char char_uplo = arg.uplo;
     char char_diag = arg.diag;
